site_server/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Django==4.0
django-perms==2.1.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
Faker==13.2.0
idna==3.3
numpy==1.21.4
pandas==1.3.5
pika==0.9.5
Pillow==8.4.0
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
